Changed Overall Appearance
I added a ramp to the 6th star to make all starts accessible. Then, I added various color materials to all the cubes and ramps to better fit the theme of the game scene. Lastly, I changed the skybox and ground to be more "space-like" in the appearance along with some sci-fi music.